User input test.
To be of practical value, an application that solicits user input must also include a mechanism for handling exceptions, e.g. invalid user input.  For example, if the user enters a letter instead of a number.  I've put this operation in Main while putting all the computational code in its own function.